Linux/OSINT_tool/requirements.txt
altgraph==0.17.4
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
av==12.1.0
beautifulsoup4==4.12.3
billiard==4.2.0
binaryornot==0.4.4
blinker==1.8.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-allauth==0.61.1
django-countries==7.5.1
django-crispy-forms==2.1
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
django-storages==1.14.2
dnspython==2.6.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
Flask-HTTPAuth==4.8.0
flatbuffers==24.3.25
flet==0.23.2
flet-core==0.23.2
flet-runtime==0.23.2
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
greenlet==3.0.3
grpcio==1.65.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
influxdb==5.3.2
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kafka-python==2.0.2
kiwisolver==1.4.5
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.82
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
phonenumbers==8.13.37
pi==0.1.2
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodomex==3.20.0
pydantic==1.10.17
pydantic_core==2.18.3
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.1.0
pyperclip==1.8.2
PyPika==0.48.9
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
python-nmap==0.7.1
python-slugify==8.0.4
python-whois==0.9.4
python3-openid==3.2.0
pytweening==1.0.7
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.4.2
regex==2023.12.25
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rjsmin==1.2.2
rpds-py==0.20.0
rsa==4.9
scapy==2.5.0
scikit-learn==1.4.2
scipy==1.13.0
Send2Trash==1.8.3
setuptools==69.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
stripe==8.3.0
striprtf==0.0.26
swapper==1.3.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.18.0.post0
threadpoolctl==3.4.0
tinycss2==1.3.0
toml==0.10.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
twisted-iocpsupport==1.0.4
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.29.0
vine==5.1.0
virtualenv==20.25.0
vonage==3.14.0
vonage-jwt==1.1.1
vonage-utils==1.1.2
watchdog==4.0.0
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.42.0
whitenoise==6.6.0
wrapt==1.12.1
xlwt==1.3.0
xxhash==3.4.1
zipp==3.18.1
